backend/requirements.txt
django==4.1
djangorestframework==3.14.0
# gunicorn==20.1.0
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
python-dotenv~=0.21.0
psycopg2-binary
whitenoise
django-cors-headers
channels
daphne
django-fsm
channels-redis